+++ lib/shrine/plugins/add_metadata.rb
@@ -24,7 +24,7 @@ class Shrine
         private
 
         def _metadata_method(name)
-          self::UploadedFile.send(:define_method, name) do
+          FileMethods.send(:define_method, name) do
             metadata[name.to_s]
           end
         end
@@ -56,6 +56,10 @@ class Shrine
             io.rewind
           end
         end
+      end
+
+      module FileMethods
+        # methods will be dynamically defined here through `Shrine.add_metadata`
       end
     end